requirements.txt
yfinance==0.2.26
pandas==2.2.3
numpy==2.2.1
python-binance==1.0.16
psycopg2-binary==2.9.10